+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * API change: Added `ProvideActionArgsRequestBody.initial`.
 * API change: Renamed `ProvideActionArgsRequestBody.features` to `argFeatures`.
 * API change: Added `minValue` and `maxValue` to `DateTimeType`.
+* Added `license` and `description` to the REST API features.
 
 ## 1.15.0 (2020-03-30)